Bugfix: Info popup button on satinfo popup doesn't work (#84)
* Removed attribute 'key' from satinfo property divs

Presumably, the 'key' attribute was added to these divs either for
performance reasons or accidentally. Either way, it's not compatible
with the way the description text popups of the buttons are
programmatically set to appear. This causes rendering issues so I
removed them.

Edit: added key attributes to span elements as well as to the epoch