Refactored generic filter code (#148)
Changed the code of some of the six main filters such that the code is
now more versatile with changes from the backend. For example, if the
backend decides to add a filter it won't automatically break the
front-end. This is because instead of using indices to select specific
fetched filter-categories from the back-end, it now uses a map to always
select the correct filter-category.